Moved AuthForm.tsx to /components
