+++ src/main.tsx
@@ -1,6 +1,6 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import AuthForm from './AuthForm.tsx'
+import AuthForm from './components/AuthForm.tsx'
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
     <AuthForm />
